// src/protocol/generated/accounts/Pool.ts
export type PoolArgs = {
  tag: number
  bump: number
  hdrPad: number[] /* size: 6 */
  authority: web3.PublicKey
  mintA: web3.PublicKey
  mintB: web3.PublicKey
  vaultA: web3.PublicKey
  vaultB: web3.PublicKey
  auditorElgamalPubkey: number[] /* size: 32 */
  reserveA: beet.bignum
  reserveB: beet.bignum
  currentBatchId: beet.bignum
  lastSettledBatchId: beet.bignum
  batchWindowSlots: number
  batchSize: number
  feeBps: number
  lpMint: web3.PublicKey
  lpEscrowShielded: web3.PublicKey
  lpSupply: beet.bignum
  genesisSlot: beet.bignum
  cumulativeVolumeA: number[] /* size: 16 */
  cumulativeVolumeB: number[] /* size: 16 */
  depositVaultA: web3.PublicKey
  depositVaultB: web3.PublicKey
  reserved: number[] /* size: 24 */
}
// ...
export const poolBeet = new beet.BeetStruct<
    Pool,
    PoolArgs
>(
  [
    
    ['tag', beet.u8],
    ['bump', beet.u8],
    ['hdrPad', beet.uniformFixedSizeArray(beet.u8, 6)],
    ['authority', beetSolana.publicKey],
    ['mintA', beetSolana.publicKey],
    ['mintB', beetSolana.publicKey],
    ['vaultA', beetSolana.publicKey],
    ['vaultB', beetSolana.publicKey],
    ['auditorElgamalPubkey', beet.uniformFixedSizeArray(beet.u8, 32)],
    ['reserveA', beet.u64],
    ['reserveB', beet.u64],
    ['currentBatchId', beet.u64],
    ['lastSettledBatchId', beet.u64],
    ['batchWindowSlots', beet.u32],
    ['batchSize', beet.u16],
    ['feeBps', beet.u16],
    ['lpMint', beetSolana.publicKey],
    ['lpEscrowShielded', beetSolana.publicKey],
    ['lpSupply', beet.u64],
    ['genesisSlot', beet.u64],
    ['cumulativeVolumeA', beet.uniformFixedSizeArray(beet.u8, 16)],
    ['cumulativeVolumeB', beet.uniformFixedSizeArray(beet.u8, 16)],
    ['depositVaultA', beetSolana.publicKey],
    ['depositVaultB', beetSolana.publicKey],
    ['reserved', beet.uniformFixedSizeArray(beet.u8, 24)]
  ],
  Pool.fromArgs,
  'Pool'
)